pyproject.toml
[tool.poetry]
name = "pluto-pizza-py"
version = "0.1.0"
description = ""
authors = ["Jonathan Osmond"]
readme = "README.md"
packages = [{include = "pluto_pizza_py"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.92.0"}
sqlalchemy = "^2.0.4"
alembic = "^1.9.4"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
factory-boy = "^3.2.1"
psycopg2-binary = "^2.9.5"
types-python-jose = "^3.3.4.5"
types-passlib = "^1.7.7.10"
pydantic = "^1.10.6"
types-factory-boy = "^0.4.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"
prospector = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
jobs=0

[tool.mypy]
exclude = "(migrations|factories)/"